Move HashingFn to io/common, switch to better hash
